==> 1_Relevant_Libraries_Installation.ipynb <==
# Asset Portfolio Management using Deep Reinforcement Learning
---

## 1.0 Load and Install Relevant Libraries and Packages
----

* We install the relevant Libraries using the FinRL Library
* We further check that all the required Libraries are installed and if not we use !pip install to install them

from importlib.metadata import distributions

installed_packages = {dist.metadata['Name'] for dist in distributions()}
print(installed_packages)
from importlib.metadata import version

pip install pyfolio

pip install finrl

----
### 1.2. Check if the additional Required packages are present. If not, install them

* Yahoo Finance API
* pandas
* numpy
* matplotlib
* stockstats
* OpenAI gym
* stable-baselines
* tensorflow
* pyfolio
* ta
* PyPortfolioOpt

required = {'yfinance', 'pandas','numpy', 'matplotlib', 'stockstats','stable-baselines',
            'gym','tensorflow','pyfolio', 'ta', 'PyPortfolioOpt'}
missing = required - installed_packages
if missing:
    !pip install yfinance
    !pip install pandas
    !pip install numpy
    !pip install matplotlib
    !pip install stockstats
    !pip install gym
    !pip install stable-baselines3[extra]
    !pip install tensorflow
    !pip install git+https://github.com/quantopian/pyfolio
    !pip install ta
    !pip install PyPortfolioOpt
    !pip install pandas_datareader
