==> 2_Data_Download.ipynb <==
# Asset Portfolio Management using Deep Reinforcement Learning
----

## 2.0 Download the Data for the Analysis
---
* Data is Downloaded from Yahoo Finance using an API in the FinRL Library
* The data is downloaded and the cleaned data are saved in csv files for use

---


### 2.1 Import Relevant Packages

import pandas as pd
import numpy as np
import my_config
import matplotlib.pylab as plt

import yfinance as yf
from pandas_datareader import data as pdr

%matplotlib inline

### 2.2 Create Folders

import os
if not os.path.exists("./" + my_config.DATA_SAVE_DIR):
    os.makedirs("./" + my_config.DATA_SAVE_DIR)
if not os.path.exists("./" + my_config.TRAINED_MODEL_DIR):
    os.makedirs("./" + my_config.TRAINED_MODEL_DIR)
if not os.path.exists("./" + my_config.TENSORBOARD_LOG_DIR):
    os.makedirs("./" + my_config.TENSORBOARD_LOG_DIR)
if not os.path.exists("./" + my_config.RESULTS_DIR):
    os.makedirs("./" + my_config.RESULTS_DIR)

### 2.3 Download the stock data

ticker_list = my_config.TICKERS

# öljy data
oil_ticker = "BZ=F"  # Yahoo Finance ticker for Brent Crude Oil Futures
oildata = yf.download(oil_ticker, start="2014-12-08", end="2025-1-1", interval="1d")
oildata = oildata.reset_index()
oildata = oildata[["Date", "Close"]]
oildata["oil_change"] = ((oildata["Close"] / oildata["Close"].shift(1)) - 1)
oildata.head()

if isinstance(oildata.columns, pd.MultiIndex):  # Check if multi-level columns exist
    oildata.columns = oildata.columns.droplevel(0)  # Flatten the columns

oildatanew = oildata.drop(columns=oildata.columns[1])  # Drop the middle column
oildatanew.columns = ["Date", "oil_change"]  # Rename the columns
oildatanew.head()

#10 vuoden bondin data
yield_ticker = "^TNX"  # Yahoo Finance ticker for 10-Year Treasury Yield
yielddata = yf.download(yield_ticker, start="2014-12-08", end="2025-1-1", interval="1d")
yielddata = yielddata.reset_index()
yielddata = yielddata[["Date", "Close"]]
yielddata["yield_change"] = ((yielddata["Close"] / yielddata["Close"].shift(1)) - 1)
yielddata.head()

if isinstance(yielddata.columns, pd.MultiIndex):  # Check if multi-level columns exist
    yielddata.columns = yielddata.columns.droplevel(0)
yielddatanew = yielddata.drop(columns=yielddata.columns[1])
yielddatanew.columns = ["Date", "yield_change"]
yielddatanew

#yhdistely
merged_macrodata = pd.merge(oildatanew, yielddatanew, on="Date", how="inner")
merged_macrodata.head()

merged_macrodata.to_csv('datasets/merged_data.csv', index=False)

import yfinance as yf

# Assuming `ticker_list` contains the list of Dow Jones tickers
ticker = ticker_list[0]  # First ticker as an example

# osakkeiden lataus
data = yf.download(ticker_list, start="2014-12-08", end="2025-1-1", group_by='ticker')

data = data.stack(level=0).reset_index()

print(data.columns)
data.columns = ['Date', 'Ticker', 'Open', 'High', 'Low', 'Close', 'Adj Close', 'Volume']

data = data.drop(columns=['Close'])

print(data.head())

data.head()

data['Ticker'] = data['Ticker'].str.replace('.HE', '', regex=False)

data.head(10)

unique_dates = data[["Date"]].drop_duplicates()  # Get unique dates from stock data
aligned_macrodata = pd.merge(unique_dates, merged_macrodata, on="Date", how="left")  # Merge with macrodata
aligned_macrodata

expanded_macrodata = aligned_macrodata.loc[aligned_macrodata.index.repeat(10)].reset_index(drop=True)

merged_data = data.copy()
merged_data["oil_change"] = expanded_macrodata["oil_change"]
merged_data["yield_change"] = expanded_macrodata["yield_change"]

merged_data  # Show the first 20 rows to check the alignment


merged_data.isna().sum()

merged_data=merged_data.dropna()
merged_data

print(merged_data.columns)

merged_data.isna().sum()

merged_data.columns=['date','tic','open','high','low','close','volume', 'oil_chg', 'yield_chg']

merged_data.columns

no_datasets = []
for i in ticker_list:
    no_data_points = merged_data[merged_data['tic']==i].shape[0]
    no_datasets.append((i,no_data_points))
    data_points_df = pd.DataFrame(no_datasets)

data_points_df

### 2.5 Save the Data to csv

merged_data.to_csv('datasets/merged_data.csv', index=False)

### 2.6 Create a Dataset for the Close Prices

df_prices = pd.read_csv('./datasets/merged_data.csv')

df_prices = df_prices.reset_index().set_index(['tic', 'date']).sort_index()

tic_list = list(set([i for i,j in df_prices.index]))
tic_list

df_prices = df_prices.reset_index().set_index(['tic', 'date']).sort_index()

df_close = pd.DataFrame()

for ticker in tic_list:
    series = df_prices.xs(ticker).close
    df_close[ticker] = series

df_close = df_close.reset_index()
df_close = df_close[['date'] + sorted(df_close.columns.drop('date'))]

df_close

df_close.describe().T


df_close.to_csv('datasets/close_prices.csv', index=False)

df_close
