==> 3_Feature_Engineering_and_Preprocessing.ipynb <==
# Asset Portfolio Management using Deep Reinforcement Learning
---

## 3.0 Feature Engineering and Data Preprocessing
---
We perform feature engineering and data preprocessing by:
* Adding Technical Indicators to the data. The technical inicators are used as inputs in the training of our Reinforcement Learning Model

### 3.1 Import Relevant Libraries

from ta import add_all_ta_features

#python env
import pandas as pd
from ta import add_all_ta_features
import numpy as np
from ta.utils import dropna
import my_config
from ta import volatility, trend, momentum

### 3.2 Load the Data from the csv Files

data = pd.read_csv('./datasets/merged_data.csv')
prices_data = pd.read_csv('./datasets/close_prices.csv')

data.head()

data = data.sort_values(by=['tic', 'date']).reset_index(drop=True)

data.head()

data.tic.unique()

### 3.3 Add Technical Indicators
---
We define a function to add technical indicators to the dataset by making use of the ta library
* Volatility Average True Range (ATR)
* Trend Fast Simple Moving Average (SMA)
* Momentum Relative Strength Index (RSI)

#VOLATILITY
# Average True Range (ATR)
#from ta.volatility import AverageTrueRange
#help(AverageTrueRange)

#TREND
# Exponential Moving Average (EMA)
#from ta.trend import EMAIndicator
#help(EMAIndicator)

#TREND
#Relative Strength Index (RSI)
#from ta.momentum import RSIIndicator
#help(RSIIndicator)

def calculate_indicators(group):
    # ATR
    group['atr'] = volatility.AverageTrueRange(
        high=group['high'], low=group['low'], close=group['close']
    ).average_true_range()

    # SMA
    group['sma'] = trend.SMAIndicator(
        close=group['close'], window=14
    ).sma_indicator()

    # RSI
    group['rsi'] = momentum.RSIIndicator(
        close=group['close'], window=14
    ).rsi()

    return group

data = data.groupby('tic').apply(calculate_indicators)
data = data.reset_index(drop=True)
data.head(17)

data = data.sort_values(by=['date', 'tic']).reset_index(drop=True)
data.head()

ticker_counts = data.groupby('tic').size()
print(ticker_counts)
data.isna().sum()

data=data.dropna()
columns = [col for col in data.columns if col not in ['oil_chg', 'yield_chg']] + ['oil_chg', 'yield_chg']
data = data[columns]
data.head()

data.to_csv('./datasets/data_with_indicators.csv', index=False)

indicator_list = list(data.columns)[7:]
indicator_list

df=data

df=df.reset_index(drop=True)

%store df

df
