==> 4_Split_Data.ipynb <==
# Asset Portfolio Management using Deep Reinforcement Learning
---

#### 4.0 Data Split
---

We will split both the close prices and the whole dataset into train and test (trade) data.

We will use 80% of the data for training and then test on the remaining 20%.

### 4.1 Import Relevant Libraries

#conda env base
import pandas as pd
import my_config
import numpy as np
import ta
from ta import add_all_ta_features
from ta.utils import dropna

### 4.2 Load the data

%store -r df
%store -r ticker_list
df_close = pd.read_csv('./datasets/close_prices.csv')

df

df.columns

df_close

start_date = df['date'].min()

df_close = df_close[df_close['date'] >= start_date]

df_close = df_close.reset_index(drop=True)

### 4.3 Split the Data

train_start_date = "2015-01-01"                        # Start of the training data
train_end_date = "2022-12-31"        # End of the training data
test_start_date = "2023-01-01"            # Start of the testing data
test_end_date = "2024-12-31"                          # End of the testing data

print(f"Training Data: from {train_start_date} to {train_end_date}")
print(f"Testing Data: from {test_start_date} to {test_end_date}")

train_data = df[(df['date'] >= train_start_date) & (df['date'] <= train_end_date)]
test_data = df[(df['date'] >= test_start_date) & (df['date'] <= test_end_date)]

train_close = df_close[(df_close['date'] >= train_start_date) & (df_close['date'] <= train_end_date)]
test_close = df_close[(df_close['date'] >= test_start_date) & (df_close['date'] <= test_end_date)]

print(f"Main training data size: {len(train_data)} rows")
print(f"Main testing data size: {len(test_data)} rows")
print(f"Close training data size: {len(train_close)} rows")
print(f"Close testing data size: {len(test_close)} rows")

### 4.4 Store the Dataframes

train_data

test_data

train_data=train_data.copy()
test_data=test_data.copy()

train_close=train_close.copy()
test_close=test_close.copy()

train_data.to_csv('datasets/train_data.csv', index=False)

test_data.to_csv('datasets/test_data.csv', index=False)

train_close.to_csv('datasets/train_close.csv', index=False)

test_close.to_csv('datasets/test_close.csv', index=False)
